Divide the report generation in two phases: prepare and generate
Having two phases is useful when you want to generate reports in parallel. The prepare phase is fast and can require information from the session. It isn't a good idea to request the session from a different thread as it can cause all sort of weird problems. The generation phase uses the prepared data and generates the report. This second phase can be time consuming and is a good candidate to run in a background thread.